feat: Introduce Router Service
Part of #7
