+++ include/Router.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Archict\Router;
+
+use Archict\Brick\Service;
+
+#[Service]
+final readonly class Router
+{
+}
